Move hot reload-related classes into separate folder
