+++ src/RealEngine/program/CMakeLists.txt
@@ -1,7 +1,6 @@
 real_target_sources(RealEngine
     PUBLIC
         CommandLineArguments.hpp    
-        HotReloadInitInfo.hpp       
         MainProgram.hpp             MainProgram.cpp
         Synchronizer.hpp            Synchronizer.cpp
 )
