+++ src/RealEngine/renderer/CMakeLists.txt
@@ -3,7 +3,6 @@ real_target_sources(RealEngine
         Allocator.hpp               
         DeletionQueue.hpp           DeletionQueue.cpp
         ObjectUsingVulkan.hpp       
-        PipelineHotLoader.hpp       PipelineHotLoader.cpp
         VulkanRenderer.hpp          VulkanRenderer.cpp
     PRIVATE
         DebugMessageHandler.hpp     DebugMessageHandler.cpp
