+++ src/RealEngine/resources/CMakeLists.txt
@@ -1,4 +1,6 @@
-﻿real_target_sources(RealEngine
+add_subdirectory(hot_reload)
+
+real_target_sources(RealEngine
     PUBLIC
         FileIO.hpp                  FileIO.cpp
         PackageConstants.hpp
