+++ src/RealEngine/resources/hot_reload/CMakeLists.txt
@@ -1,0 +1,5 @@
+real_target_sources(RealEngine
+    PUBLIC
+        HotReloadInitInfo.hpp       
+        PipelineHotLoader.hpp       PipelineHotLoader.cpp
+)
