+++ src/RealEngine/window/Window.hpp
@@ -8,8 +8,8 @@
 #include <SDL_video.h>
 #include <glm/vec2.hpp>
 
-#include <RealEngine/program/HotReloadInitInfo.hpp>
 #include <RealEngine/renderer/VulkanRenderer.hpp>
+#include <RealEngine/resources/hot_reload/HotReloadInitInfo.hpp>
 #include <RealEngine/utility/UniqueCPtr.hpp>
 #include <RealEngine/window/WindowSettings.hpp>
 #include <RealEngine/window/WindowSubsystems.hpp>
